Reverted IOError capture and expanded tests.
Authenticator.control() raises ValueError or KeyError on certain conditions,
it was changed to remove one of those conditions in order to capture potential
IOErrors. This was due to a misunderstanding of what the method does, which is
why I am reverting it.

The issue of the IOError is still a question. I have not added any code to
handle it yet because I do not see an obvious Right Thing to do except for
letting the error propagate up the stack.

I have added more tests for control() because I had neglected to test a couple
of the error conditions.